+++ Source/Engine/AGS.Engine/Objects/Characters/Walking/Following/AGSFollowComponent.cs
@@ -107,8 +107,8 @@ namespace AGS.Engine
 
 		private PointF wanderOff()
 		{
-			float x = (float)MathUtils.Random().Next(0, _game.Settings.VirtualResolution.Width);
-			float y = (float)MathUtils.Random ().Next (0, _game.Settings.VirtualResolution.Height);
+			float x = MathUtils.Random().Next(0, _game.Settings.VirtualResolution.Width);
+			float y = MathUtils.Random ().Next (0, _game.Settings.VirtualResolution.Height);
 			return new PointF(x, y);
 		}
